updated install/uninstall script

+++ uninstall.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
 #
-#  This file is part of Etherwall
-#  Copyright (C) Agus Bimantoro <l0g.bima@gmail.com>
-#  This program is published under a GPLv3 license
+# This file is part of Etherwall
+# Copyright (C) Agus Bimantoro <l0g.bima@gmail.com>
+# This program is published under a GPLv3 license
 
-who_uninstall=`whoami`
+if [ "$(whoami)" != "root" ]; then
+  echo "Error: you need to be root to remove etherwall"
+  exit 1
+fi
 
-function main {
-echo -e "Uninstalling etherwall:\n"
+echo -n "Are you sure you want to remove etherwall?[y/n] "
+read cek
+if [ "$cek" != "y" -a "$cek" != "Y" ]; then
+  echo "Uninstallation aborted."
+  exit 1
+fi
+
+if [ ! -d "/opt/etherwall" ]; then
+  echo "Error: etherwall is not installed on your system."
+  exit 1
+fi
 
 echo "Removing etherwall installation directory..."
 rm -r /opt/etherwall
@@ -16,37 +28,12 @@ echo "Removing etherwall configuration directory..."
 rm -r /etc/etherwall 2> /dev/null
 
 echo "Removing etherwall log directory..."
-rm /var/log/etherwall/etherwall.log
+rm -r /var/log/etherwall
 
-echo "Removing etherwall script program..."
-rm /sbin/etherwall 2> /dev/null
+echo "Removing etherwall scripts..."
+rm /sbin/{etherwall,ethwconsole} 2> /dev/null
 
-echo "Removing ethwconsole script program..."
-rm /sbin/ethwconsole 2> /dev/null
+echo "Removing etherwall manual pages..."
+  rm /usr/share/man/man8/{etherwall.8.gz,etherwall-id.8.gz,ethwconsole.8.gz,ethwconsole-id.8.gz} 2> /dev/null
 
-echo "Removing etherwall manual page..."
-rm /usr/share/man/man8/etherwall.8.gz 2> /dev/null
-rm /usr/share/man/man8/etherwall-id.8.gz 2> /dev/null
-rm /usr/share/man/man8/ethwconsole.8.gz 2> /dev/null
-rm /usr/share/man/man8/ethwconsole-id.8.gz 2> /dev/null
-
-echo -e "\nUninstallation finished."
-}
-
-if [ "$who_uninstall" == "root" ]; then
-        echo -n -e "Are you sure you want to remove/uninstall etherwall [Y/n]? "
-	read persetujuan
-	if [ "$persetujuan" == "Y" ] || [ "$persetujuan" == "y" ]; then		
-		if [ -d "/opt/etherwall/" ]; then 
-			main
-		else
-			echo -e "\nError: Etherwall program is not installed on your system"
-			exit 1
-		fi
-	else
-		exit 0
-	fi
-else
-	echo -e "Error: Cannot remove/uninstall etherwall,\n       it may require superuser privileges (eg. root)."
-	exit 1
-fi
+echo "All done..."
